+++ NEWS.md
@@ -1,3 +1,5 @@
+# wikkitidy (development version)
+
 # wikkitidy 0.1.12
 
 * `tidyr` moved from Imports to Suggests