+++ NEWS.md
@@ -1,5 +1,7 @@
 # wikkitidy (development version)
 
+* New `gracefully()` function added to allow graceful failure of http requests, as per CRAN policy. All API calls to Wikipedia query modules are now wrapped in `gracefully()` in the examples.
+
 # wikkitidy 0.1.12
 
 * `tidyr` moved from Imports to Suggests